feat: introduce core implementations for 1D basis functions
- Added new modules for 1D basis function evaluations, including Bernstein, cardinal B-spline, Lagrange, and Legendre basis functions.
- Implemented core functions with input validation and output management, enhancing performance through Numba compilation.
- Established helper functions for tabulating basis functions, ensuring consistent behavior and improved maintainability.
- Updated documentation to reflect the new implementations and their usage.
