+++ src/pantr/change_basis.py
@@ -11,7 +11,7 @@ from collections.abc import Callable
 import numpy as np
 import numpy.typing as npt
 
-from ._basis_impl import _get_lagrange_points
+from ._basis_lagrange import _get_lagrange_points
 from ._basis_utils import _validate_out_array_multidimensional
 from .basis import LagrangeVariant, tabulate_Bernstein_basis_1D, tabulate_cardinal_Bspline_basis_1D
 from .quad import get_gauss_legendre_quadrature_1D
